Fix/gcms uat enhancements EGCMS-110 (#14142)
* fix: revert blade filters in reseller accounts for consistency

* test: add e2e for clear button

* fix: lint error

* fix: uat bug & ui fixes

* fix: Pagination changes

* refactor(ResellerAccounts): replace DataTable with  & add total count in table TableBody

* feat: add gcms feature flag condition

* fix: back button, global cart , order create button

* Fix: Filter brand transaction with account id [EGCMS-116] (#14117)

* adding account id filter for transactions

* fixing account id filter

* small refactor

* fixing PR comments

* fix: Order Programs, permanent card footer

* fix: Resellers table padding

* fix: Fix create programs denominations modal changes and UI fixes

* add order id filter (#14129)

* fix: Fix create programs cart screens

* fix: Fix discount percentage

* fix: Fix tests

* fix: Fix flaky tests

* skip transactions

* fix: Fix flaky GCMS Order Create test tests

* fix: Fix flaky Order tests

* fix: Fix e2e tests

* fix: Fix e2e tests

* feat: Add Global Cart and Order button

* feat: Add Cancel Order functionality, multiple orders

* fix: lint errors

* fix: ts errors

* fix: e2e tests

* fix: Prod Test Base URL path

* fix: Mobile UI Changes

* fix: Lint

---------

+++ web/js/merchant/views/GCMS/Orders/OrderCartSummarySection.tsx
@@ -84,7 +84,7 @@ const OrderCartSummarySection = () => {
                               <Box>
                                 <Text weight="bold">
                                   {/* @ts-expect-error parseInt Number */}
-                                  {parseInt(program?.default_discount / 100, 10)}%
+                                  {parseFloat(program?.default_discount / 100, 10)}%
                                 </Text>
                               </Box>
                             </Box>
